modify sns  upload ses

+++ streamNS.h
@@ -20,13 +20,13 @@ public:
 	int size;
 	string outputFile;
 
-	map<uint64_t,int> VShash; //点ID的hash值，点ID
-	map<int,int> VSdegree;//点ID，点ID的度
-	multimap<int,int> ES;
+	map<uint64_t,uint64_t> VShash; //点ID的hash值，点ID
+	map<uint64_t,int> VSdegree;//点ID，点ID的度
+	multimap<uint64_t,uint64_t> ES;
 
 	void process();
-	uint64_t hash64( int , int );
-	void delEdge( int , int );
+	uint64_t hash64( uint64_t , int );
+	void delEdge( uint64_t , int );
 	void saveSampleGraph();
 };
